+++ tests/basics/explicit_nesting/src/foo.rb
