+++ tests/basics/explicit_nesting/src/foo/bar/baz.rb
