+++ tests/basics/implicit_nesting/src/foo/bar/baz.rb
