+++ tests/basics/implicit_nesting/src/foo/bar/qux.rb
