+++ tests/basics/implicit_nesting/test.rb
