+++ tests/basics/related_top_level_constants/src/bar.rb
