+++ tests/basics/related_top_level_constants/src/baz.rb
