+++ tests/basics/related_top_level_constants/src/foo.rb
