+++ tests/basics/related_top_level_constants/test.rb
