+++ tests/basics/unrelated_top_level_constants/src/bar.rb
