+++ tests/basics/unrelated_top_level_constants/test.rb
