+++ tests/edge_cases/redundant_nesting/src/this/nests/really/really/really/deeply.rb
