+++ tests/edge_cases/redundant_nesting/test.rb
