+++ tests/errors/directory_shadows_existing_constant/src/array/immutable.rb
