+++ tests/errors/directory_shadows_existing_constant/test.rb
