+++ tests/errors/file_breaks_convention/src/bar.rb
