+++ tests/errors/file_breaks_convention/src/foo.rb
