+++ tests/errors/file_breaks_convention/test.rb
