+++ tests/errors/file_references_unknown_constant/src/foo.rb
