+++ tests/errors/file_references_unknown_constant/test.rb
