+++ tests/errors/file_shadows_existing_constant/src/array.rb
