+++ tests/errors/file_shadows_existing_constant/test.rb
