+++ tests/features/custom_inflections/src/models/dwarf_rogue.rb
