+++ tests/features/custom_inflections/src/models/elf_archer.rb
