+++ tests/features/custom_inflections/src/models/half_elf_archer.rb
