+++ tests/features/custom_inflections/src/ui/button.rb
