+++ tests/features/custom_inflections/src/ui/link_to_url.rb
