+++ tests/features/custom_inflections/test.rb
