+++ tests/features/ignored_files/src/entities/player.rb
