+++ tests/features/ignored_files/src/entity.rb
