+++ tests/features/ignored_files/src/ignored.rb
