+++ tests/features/ignored_files/src/scene.rb
