+++ tests/features/ignored_files/test.rb
