+++ tests/features/multiple_roots/src/components/clickable.rb
