+++ tests/features/multiple_roots/src/components/hoverable.rb
