+++ tests/features/multiple_roots/src/systems/mouse_system.rb
