+++ tests/features/multiple_roots/test.rb
