+++ tests/features/nested_roots/src/lib/entities/car.rb
