+++ tests/features/nested_roots/src/lib/entities/static/billboard.rb
