+++ tests/features/nested_roots/src/lib/scene.rb
