+++ tests/features/nested_roots/src/lib/stages/stage_two.rb
